Moved suggest id and add comment buttons out of activity tab, added stickybar, PhotoScroll buttons, back button
