+++ src/navigation/BottomTabNavigator/CustomTabBar.js
@@ -69,7 +69,7 @@ const CustomTabBar = ( { state, descriptors, navigation }: Props ): Node => {
     </View>
   );
 
-  const footerHeight = Platform.OS === "ios" ? "h-80px]" : "h-20";
+  const footerHeight = Platform.OS === "ios" ? "h-[80px]" : "h-20";
 
   // Hacky solution but is required to show ContextHeader shadow in PhotoGallery
   // when PhotoGallery is hoisted to stack navigator, the header is rendered first
@@ -91,7 +91,7 @@ const CustomTabBar = ( { state, descriptors, navigation }: Props ): Node => {
   return (
     <View
       className={classNames(
-        "flex flex-row absolute bottom-0 bg-white w-full justify-evenly items-center p-1",
+        "flex flex-row absolute bottom-0 bg-white w-full justify-evenly items-center p-1 m-0",
         { "pb-5": Platform.OS === "ios" },
         footerHeight
       )}
